backend/app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.66
botocore==1.42.66
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cryptography==46.0.5
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docx==0.2.4
dotenv==0.9.9
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.133.0
fastapi-sa-orm-filter==0.2.5
frozenlist==1.8.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.15
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
langsmith==0.7.6
librt==0.8.1
limits==5.8.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy==1.20.0
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
openai==2.23.0
openpyxl==3.1.5
openpyxl-stubs==0.1.25
orjson==3.11.7
packaging==26.0
pathspec==1.0.4
pgvector==0.4.2
pillow==12.2.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==7.34.1
psycopg2==2.9.11
pwdlib==0.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
pypdf==6.9.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.2.1
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlalchemy-to-pydantic==0.0.8
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
